requirements.txt
fsspec
pyworld
matplotlib
soundfile
scikit-learn
scipy
tensorboard
transformers
tqdm
librosa
omegaconf
gradio==3.36.1
ruamel.yaml
resampy
numpy==1.24
chardet
faiss-cpu==1.7.4
